+++ NEWS.md
@@ -4,6 +4,7 @@
 ### New functions / parameters
 - Added R2 for bayesian regressions.
 - Added `dprime` function for signal detection theory indices computation.
+- Added `crawford.test` and `mellenbergh.test` function for single-case tests.
 ### Changes